@@ -1,13 +1,11 @@
 name: app_popup_menu
-description: Your app's popup menu. It's adaptive and dynamic.
+description: A more modular popup menu.
+version: 1.0.0
 homepage: https://github.com/AndriousSolutions/app_popup_menu
 repository: https://github.com/AndriousSolutions/app_popup_menu
 
-version: 0.6.0+5
-
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
