@@ -1,13 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import '../example/lib/main.dart';
+import '../example/main.dart';
 
 void main() {
   test('adds one to input values', () {
-    final calculator = MyApp();
-    // expect(calculator.addOne(2), 3);
-    // expect(calculator.addOne(-7), -6);
-    // expect(calculator.addOne(0), 1);
-    // expect(() => calculator.addOne(null), throwsNoSuchMethodError);
+    final app = const MyApp();
+
   });
 }
